requirements.txt
# Зависимости для Windows Optimization Bot
# Точные версии для обеспечения совместимости

# Telegram Bot API
pyTelegramBotAPI==4.15.0

# Anthropic API (для Claude)
anthropic

# Асинхронный клиент HTTP
aiohttp==3.9.3
asyncio==3.4.3

# Работа с изображениями
pillow==10.2.0

# Переменные окружения
python-dotenv==1.0.0

# HTTP-запросы
requests==2.31.0

# Логирование и диагностика
colorlog==6.8.0

# Flask (опционально, если нужен веб-сервер для проверки работоспособности)
# flask==2.3.3